+++ CHANGELOG.md
@@ -110,6 +110,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Removed
 
+- Remove the language selector from the selector widget (#4262)
 - Remove the country selector from the selector widget (#4261)
 - Drop the `defaultEventVipsFeGroupID` configuration option (#4260)
 - Drop the `allowCsvExportForVips` configuration option (#4233, #4259)